Seperate ChunkRenderer into ChunkFactory and ChunkRenderer
