+++ source/marchingcubes/CMakeLists.txt
@@ -38,6 +38,7 @@ set(source_path "${CMAKE_CURRENT_SOURCE_DIR}/")
 set(sources
     ${source_path}/AddChunksStage.cpp
     ${source_path}/Chunk.cpp
+    ${source_path}/ChunkFactory.cpp
     ${source_path}/ChunkRenderer.cpp
     ${source_path}/LookUpData.cpp
     ${source_path}/ManageChunksStage.cpp
@@ -50,6 +51,7 @@ set(sources
 set(api_includes
     ${include_path}/AddChunksStage.h
     ${include_path}/Chunk.h
+    ${include_path}/ChunkFactory.h
     ${include_path}/ChunkRenderer.h
     ${include_path}/LookUpData.h
     ${include_path}/ManageChunksStage.h
