+++ source/marchingcubes/ChunkRenderer.h
@@ -35,56 +35,19 @@ public:
 
     void setTransform(glm::mat4x4 transform);
 
-    void generateDensities(Chunk * chunk);
-    void generateList(Chunk * chunk);
-    void generateMesh(Chunk * chunk);
-
 	void updateTexture(bool useMipMap);
-
-    globjects::Program * densityGenerationProgram() const { return m_densityGenerationProgram.get(); }
 
 protected:
     void setupProgram();
     void setupRendering();
-    void setupDensityGeneration();
-    void setupListGeneration();
-    void setupMeshGeneration();
-    void setupNoiseTextures();
-    globjects::ref_ptr<globjects::Texture> setupNoiseTexture(glm::vec3 offset);
 
 protected:
-    globjects::ref_ptr<globjects::Buffer> m_positions;
     gl::GLint m_transformLocation;
     gl::GLint m_offsetLocation;
 
-    globjects::ref_ptr<globjects::VertexArray> m_densityPositionVao;
-    globjects::ref_ptr<globjects::VertexArray> m_meshVao;
-    globjects::ref_ptr<globjects::VertexArray> m_listVao;
-
-    globjects::ref_ptr<globjects::Buffer> m_densityPositions;
-    gl::GLsizei m_densityPositionsSize;
-
-    globjects::ref_ptr<globjects::Buffer> m_edgeConnectList;
-
-    globjects::ref_ptr<globjects::TransformFeedback> m_densityGenerationTransformFeedback;
-    globjects::ref_ptr<globjects::Program> m_densityGenerationProgram;
-
-    globjects::ref_ptr<globjects::TransformFeedback> m_listGenerationTransformFeedback;
-    globjects::ref_ptr<globjects::Program> m_listGenerationProgram;
-
-    globjects::ref_ptr<globjects::TransformFeedback> m_meshGenerationTransformFeedback;
-    globjects::ref_ptr<globjects::Program> m_meshGenerationProgram;
-
     globjects::ref_ptr<globjects::Program> m_renderProgram;
 
-    gl::GLsizei m_positionsSize;
-
     glm::mat4x4 m_transform;
-
-    globjects::ref_ptr<globjects::Texture> m_noiseTexture1;
-    globjects::ref_ptr<globjects::Texture> m_noiseTexture2;
-    globjects::ref_ptr<globjects::Texture> m_noiseTexture3;
-    globjects::ref_ptr<globjects::Texture> m_noiseTexture4;
 
 	globjects::ref_ptr<globjects::Texture> m_colorTexture;
 	globjects::ref_ptr<globjects::Texture> m_groundTexture;
